backend/requirements.txt
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
et_xmlfile==2.0.0
flasgger==0.9.7.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.1.4
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
